+++ src/main/java/project/carbonFootprint/models/dto/CalculationResponse.java
@@ -12,6 +12,6 @@ import project.users.models.dto.UserResponse;
 @Builder
 public class CalculationResponse {
     UserResponse user;
-    Double totalCo2Emitted;
+    Integer totalCo2Emitted;
     Integer totalGreenScore;
 }
